+++ CHANGELOG.md
@@ -1,2 +1,11 @@
+## Version 0.3.0
+- major overhaul, ported/remade the project in Vite+Svelte+TS
+- added syntax highlighting with highlight.js
+- added feature to export multiple yarn files from one twine story
+
+## Version 0.2.0
+- minor changes and improvements to the script
+- added logo
+
 ## Version 0.1.0
 - initial release of YarnTwiner